ENH: Simplify real esv computations
This keeps track of both the start and end of each reference
chromosome, and doesn't bother distinguishing between if the start/end
of an edge are on the same chromosome or not (it looks at the
chromosomes of the start and end position regardless).

This had a few ramifications on how certain support values for
corner-case-ish edges were calculated. I think this is a step in
the right direction, though.

Would be good to add test code for this, if I can scrounge up a combo of
publicly available + free PacBio and AmpliconArchitect data